+++ R/get_calculatedmetrics.R
@@ -81,56 +81,52 @@
 #'
 #' @return A data frame of calculated metrics and their metadata.
 #'
-#' @import stringr
-#' @importFrom utils URLencode
 #' @export
 #'
 aw_get_calculatedmetrics <- function(company_id = Sys.getenv("AW_COMPANY_ID"),
-                          rsids = NA,
-                          ownerId = NA,
-                          filterByIds = NA,
-                          toBeUsedInRsid = NA,
+                          rsids = NULL,
+                          ownerId = NULL,
+                          filterByIds = NULL,
+                          toBeUsedInRsid = NULL,
                           locale = "en_US",
-                          name = NA,
-                          tagNames = NA,
-                          favorite = NA,
-                          approved = NA,
+                          name = NULL,
+                          tagNames = NULL,
+                          favorite = NULL,
+                          approved = NULL,
                           limit = 1000,
                           page = 0,
                           sortDirection = 'DESC',
-                          sortProperty = NA,
-                          expansion = NA,
+                          sortProperty = NULL,
+                          expansion = NULL,
                           includeType = 'all',
                           debug = FALSE)
 {
-  #edit the character vectors to the string they need to be
-  if(length(rsids) > 1) {rsids = utils::URLencode(paste0(rsids, collapse = ',')) }
-  if(length(filterByIds) > 1) {filterByIds = paste0(filterByIds, collapse = ',') }
-  if(!is.na(tagNames)) {tagNames = utils::URLencode(paste0(tagNames, collapse = ',')) }
-  if(length(expansion) > 1) {expansion = paste(expansion, collapse = ',', sep = '') }
+  # Reference: https://github.com/AdobeDocs/analytics-2.0-apis/blob/master/calculatedmetrics.md
+  assertthat::assert_that(length(name) < 2, msg = "'name' is a search string and may not have length > 1\nUse 'filterByIds' to request specific segments")
 
-  #includeType is case senstative
-  includeType <- tolower(includeType)
 
-  vars <- tibble::tibble(rsids, ownerId, filterByIds, toBeUsedInRsid, locale, name, tagNames, favorite, approved,
-                         limit, page, sortDirection, sortProperty, expansion, includeType)
-  #Turn the list into a string to create the query
-  prequery <- vars %>% dplyr::select_if(~ !any(is.na(.)))
-  #remove the extra parts of the string and replace it with the query parameter breaks
-  query_param <-  paste(names(prequery), prequery, sep = '=', collapse = '&')
+  query_param <- list(
+    rsids = rsids,
+    ownerId = ownerId,
+    filterByIds = filterByIds,
+    toBeUsedInRsid = toBeUsedInRsid,
+    locale = "en_US",
+    name = name,
+    tagNames = tagNames,
+    favorite = favorite,
+    approved = approved,
+    limit = 1000,
+    page = 0,
+    sortDirection = 'DESC',
+    sortProperty = sortProperty,
+    expansion = expansion,
+    includeType = 'all'
+  )
 
-  #create the url to send with the query
-  urlstructure <- paste0('calculatedmetrics?',query_param)
+  urlstructure <- paste('calculatedmetrics', format_URL_parameters(query_param), sep = "?")
 
-  #urlstructure <- 'segments?locale=en_US&filterByPublishedSegments=all&limit=1000&page=0&sortDirection=ASC&sortProperty=id&includeType=all'
   res <- aw_call_api(req_path = urlstructure[1], debug = debug, company_id = company_id)
 
-  res <- jsonlite::fromJSON(res)
-
-  #Just need the content of the returned json
-  res <- res$content
-
-  res
-
-  }
+  jsonlite::fromJSON(res)$content
+}