+++ R/get_me.R
@@ -3,8 +3,6 @@
 #' This function will quickly pull the list of company ids that you have access to.
 #'
 #' @param req_path The endpoint for that particular report
-#' @param client_id Set in environment args, or pass directly here
-#' @param client_secret Set in environment args, or pass directly here
 #'
 #' @return A data frame of company ids and company names
 #' @examples
@@ -13,18 +11,17 @@
 #' }
 #' @export
 #' @import assertthat httr
-get_me <- function(req_path = 'discovery/me',
-                   client_id = Sys.getenv("AW_CLIENT_ID"),
-                   client_secret = Sys.getenv("AW_CLIENT_SECRET")) {
-
+get_me <- function(req_path = 'discovery/me') {
     assertthat::assert_that(
-        assertthat::is.string(req_path),
-        assertthat::is.string(client_id),
-        assertthat::is.string(client_secret)
+        is.string(req_path)
     )
+
+    env_vars <- get_env_vars()
+    token_config <- get_token_config(client_id = env_vars$client_id,
+                                     client_secret = env_vars$client_secret)
+
     request_url <- sprintf("https://analytics.adobe.io/%s",
                            req_path)
-    token_config <- get_token_config(client_id = client_id, client_secret = client_secret)
 
     req <- httr::RETRY("GET",
                        url = request_url,
@@ -32,7 +29,7 @@ get_me <- function(req_path = 'discovery/me',
                        body = FALSE,
                        token_config,
                        httr::add_headers(
-                           `x-api-key` = client_id
+                           `x-api-key` = env_vars$client_id
                        ))
 
 
@@ -42,7 +39,7 @@ get_me <- function(req_path = 'discovery/me',
 
     me <- jsonlite::fromJSON(res)
 
-    return(me$imsOrgs$companies %>%
+    me$imsOrgs$companies %>%
                dplyr::bind_rows() %>%
-               dplyr::select(1:2))
+               dplyr::select(1:2)
 }